+++ src/ui/pages/page_setup.lua
@@ -18,9 +18,9 @@ sm = SetupManager()
 local function setupItemWindow()
         cui.pushContentWindow(
                 "car_setup_items_window",
-                ui.windowWidth() * 0.22 + 15 * cui.scale(),
+                ui.windowWidth() * 0.25 + 15 * cui.scale(),
                 0,
-                ui.windowWidth() * 0.56 - 30 * cui.scale(),
+                ui.windowWidth() * 0.5 - 30 * cui.scale(),
                 ui.windowHeight(),
                 function()
                         ui.setCursor(0)
@@ -54,7 +54,6 @@ local carSetupWindow = WidgetWindow("car_setup")
 carSetupWindow:addWidget("Car Setup", require("ui.widgets.setup_tab_bar"))
 carSetupWindow:addWidget("Local Setups", require("ui.widgets.local_setups"))
 carSetupWindow:addWidget("Setup Exchange", require("ui.widgets.setup_exchange"))
--- carSetupWindow:addWidget("Setup Exchange", function() setupExchange:draw() end)
 
 local saveSetupWindow = WidgetWindow("save_setup")
 saveSetupWindow:addWidget("Save Setup", require("ui.widgets.save_setup"))
@@ -70,26 +69,26 @@ function page:draw()
 
         if helpWindowDelay > os.clock() then
                 carSetupWindow:setPosition(0, 0)
-                carSetupWindow:setSize(ui.windowWidth() * 0.22, ui.windowHeight() * 0.75 - 7.5 * cui.scale())
+                carSetupWindow:setSize(ui.windowWidth() * 0.25, ui.windowHeight() * 0.75 - 7.5 * cui.scale())
                 carSetupWindow:draw()
 
                 saveSetupWindow:setPosition(0, ui.windowHeight() * 0.75 + 7.5 * cui.scale())
-                saveSetupWindow:setSize(ui.windowWidth() * 0.22, ui.windowHeight() * 0.25 - 7.5 * cui.scale())
+                saveSetupWindow:setSize(ui.windowWidth() * 0.25, ui.windowHeight() * 0.25 - 7.5 * cui.scale())
                 saveSetupWindow:draw()
         else
                 setupHelpWidget:setPosition(0, 0)
-                setupHelpWidget:setSize(ui.windowWidth() * 0.22, ui.windowHeight())
+                setupHelpWidget:setSize(ui.windowWidth() * 0.25, ui.windowHeight())
                 setupHelpWidget:draw()
         end
 
         setupItemWindow()
 
-        carStatusWindow:setPosition(ui.windowWidth() * 0.78, 0)
-        carStatusWindow:setSize(ui.windowWidth() * 0.22, ui.windowHeight() * 0.84 - 7.5 * cui.scale())
+        carStatusWindow:setPosition(ui.windowWidth() * 0.75, 0)
+        carStatusWindow:setSize(ui.windowWidth() * 0.25, ui.windowHeight() * 0.84 - 7.5 * cui.scale())
         carStatusWindow:draw()
 
-        dataLoggerWindow:setPosition(ui.windowWidth() * 0.78, ui.windowHeight() * 0.84 + 7.5 * cui.scale())
-        dataLoggerWindow:setSize(ui.windowWidth() * 0.22, ui.windowHeight() * 0.16 - 7.5 * cui.scale())
+        dataLoggerWindow:setPosition(ui.windowWidth() * 0.75, ui.windowHeight() * 0.84 + 7.5 * cui.scale())
+        dataLoggerWindow:setSize(ui.windowWidth() * 0.25, ui.windowHeight() * 0.16 - 7.5 * cui.scale())
         dataLoggerWindow:draw()
 
         return app.state.setupTab > 1 and "finalize" or "apps"